[TNT-114] feat: 회원가입 API 구현
- 몸무게 Float에서 Double로 수정
- FileUtils 이용해 이미지 파일 변환
- SignUpRequestMapper를 통해 trainer, trainee 데이터 다르게 전달

+++ domain/src/main/java/co/kr/tnt/domain/model/UserType.kt
@@ -19,7 +19,7 @@ sealed class UserType {
         override val image: String? = null,
         val birthday: LocalDate? = null,
         val age: Int? = 0,
-        val weight: Float = 0f,
+        val weight: Double = 0.0,
         val height: Int = 0,
         val ptPurpose: List<String> = emptyList(),
         val caution: String? = "",
